+++ examples/tau_vg_api.py
@@ -58,7 +58,7 @@ def _get_qc_by_sample_name(sample_sheet_lookup: SampleSheetLookup, sequencing_fi
     bam_and_vcf_by_name = {}
     for sf in sequencing_files:
         bam_file = dataclasses.replace(sf.bam_file, aligner=None)
-        vcf_file = dataclasses.replace(sf.vcf_file, variant_caller=None)
+        vcf_file = dataclasses.replace(sf.get_vcf_files()[0], variant_caller=None)
         bam_and_vcf_by_name[sf.sample_name] = (bam_file, vcf_file)
 
     qc_by_name = {}
@@ -252,7 +252,7 @@ def run_api(server, api_token, step=None):
                        fastq_r1=seq_run_path(f"0_fastq/{SAMPLE}_R1.fastq.gz"),
                        fastq_r2=seq_run_path(f"0_fastq/{SAMPLE}_R2.fastq.gz"),
                        bam_file=bam_file,
-                       vcf_file=vcf_file)
+                       vcf_files=[vcf_file])
 
     ]